+++ environment.yml
@@ -33,7 +33,6 @@ dependencies:
   - sphinx_rtd_theme
   - tornado>=6.1
   - traitsui
-  - uvloop>=0.16
   - wxPython
   - zlib
   - zict
