+++ mosaic/profile/profile.py
@@ -4,8 +4,8 @@ import sys
 import glob
 import copy
 import time
-import pickle
 import atexit
+import pickle
 import inspect
 import asyncio
 import datetime
