+++ mosaic/runtime/__init__.py
@@ -5,3 +5,4 @@ from .monitor import *
 from .node import *
 from .worker import *
 from .warehouse import *
+from .utils import *
