+++ mosaic/utils/__init__.py
@@ -5,3 +5,5 @@ from .logger import *
 from .subprocess import *
 from .utils import *
 from .change_case import *
+from .spill_buffer import *
+from .at_exit import *
