+++ mosaic/utils/event_loop.py
@@ -1,7 +1,6 @@
 
 import os
 import uuid
-import uvloop
 import asyncio
 import inspect
 import weakref
@@ -174,7 +173,6 @@ class EventLoop:
     """
 
     def __init__(self, loop=None):
-        uvloop.install()
         self._loop = loop or asyncio.new_event_loop()
         asyncio.set_event_loop(self._loop)
